_archived_/api-front/requirements.txt
azure-functions==1.20.0
azure-cosmos==4.7.0

numpy==2.0.1
